ajustando arestas

+++ modulo6/projeto-labook/src/business/FriendshipBusiness.ts
@@ -1,4 +1,5 @@
 import { FriendshipDatabase } from "../data/FriendshipDatabase";
+import { InvalidRequest } from "../error/InvalidRequest";
 import { friendshipDTO } from "../model/friendship";
 import { generateId } from "../services/generateId";
 
@@ -8,7 +9,7 @@ export class FriendshipBusiness {
             const { userId, friendId } = input
 
             if ( !userId || !friendId ) {
-                throw new Error("Todos os campos são obrigatórios!");
+                throw new InvalidRequest()
             }
 
             const id1: string = generateId()
@@ -26,17 +27,22 @@ export class FriendshipBusiness {
                 friendId
             })
         } catch (error: any) {
-            throw new Error(error.message);
+            throw new Error(error.message || error.sqlMessage);
         }
     };
 
     public deleteFriendship = async (input: friendshipDTO) => {
         try {
+            const { userId, friendId } = input
+
+            if ( !userId || !friendId ) {
+                throw new InvalidRequest()
+            }
 
             const friendshipDatabase = new FriendshipDatabase()
             await friendshipDatabase.deleteFriendship(input)
         } catch (error:any) {
-            throw new Error(error.message)
+            throw new Error(error.message || error.sqlMessage)
         }
     }
 }
